+++ src/components/spartan/STable/STable.stories.ts
@@ -34,66 +34,66 @@ export default {
     },
 };
 
-const cols = ['Name', 'Title', 'Email', 'Role', ''];
+const cols = ['Name', 'Email', 'Title', 'Role', ''];
 const rows = [
     {
         name: 'Lindsay Walton',
-        title: 'Front-end Developer',
         email: 'lindsay.walton@example.com',
+        title: 'Front-end Developer',
         role: 'Member',
     },
     {
         name: 'Jhon Connor',
-        title: 'Back-end Developer',
         email: 'jhon.connor@example.com',
+        title: 'Back-end Developer',
         role: 'Member',
     },
     {
         name: 'Jane Doe',
-        title: 'Full Stack Developer',
         email: 'jane.doe@example.com',
+        title: 'Full Stack Developer',
         role: 'Admin',
     },
     {
         name: 'John Smith',
-        title: 'Data Scientist',
         email: 'john.smith@example.com',
+        title: 'Data Scientist',
         role: 'Member',
     },
     {
         name: 'Emma Jones',
-        title: 'UX Designer',
         email: 'emma.jones@example.com',
+        title: 'UX Designer',
         role: 'Member',
     },
     {
         name: 'Robert Brown',
-        title: 'Product Manager',
         email: 'robert.brown@example.com',
+        title: 'Product Manager',
         role: 'Admin',
     },
     {
         name: 'Sophia Davis',
-        title: 'QA Engineer',
         email: 'sophia.davis@example.com',
+        title: 'QA Engineer',
         role: 'Member',
     },
     {
         name: 'Michael Miller',
-        title: 'DevOps Engineer',
         email: 'michael.miller@example.com',
+        title: 'DevOps Engineer',
         role: 'Member',
     },
     {
         name: 'Olivia Wilson',
-        title: 'Mobile Developer',
         email: 'olivia.wilson@example.com',
+        title: 'Mobile Developer',
         role: 'Member',
     },
     {
         name: 'James Moore',
-        title: 'Security Engineer',
         email: 'james.moore@example.com',
+        title: 'Security Engineer',
         role: 'Admin',
     },
 ];
@@ -132,10 +132,10 @@ export const Default = createDefault({
                 {{ row.name }}
             </STableCell>
             <STableCell>
-                {{ row.title }}
+                {{ row.email }}
             </STableCell>
             <STableCell>
-                {{ row.email }}
+                {{ row.title }}
             </STableCell>
             <STableCell>
                 <SBadge :color="row.role === 'Admin' ? 'yellow' : 'green'">{{ row.role }}</SBadge>
@@ -194,45 +194,142 @@ export const Base = createVariation({
         DocumentDuplicateIcon,
         TrashIcon,
     },
+    containerClass: 'w-fit',
     setup: () => {
         const getRoleColor = (role: string) => {
             return role === 'Admin' ? 'yellow' : 'green';
         }
 
-        return { cols, rows, getRoleColor };
+        return { cols: cols.slice(0, 2), rows: rows.slice(0, 4), getRoleColor };
     },
     template: `<STable>
     <STableHead>
         <STableHeadCell v-for="col in cols" :key="col">{{ col }}</STableHeadCell>
     </STableHead>
-
     <STableBody>
         <STableRow v-for="row in rows" :key="row.email">
-            <STableCell highlight>
+            <STableCell>
                 {{ row.name }}
             </STableCell>
             <STableCell>
-                {{ row.title }}
-            </STableCell>
-            <STableCell>
                 {{ row.email }}
-            </STableCell>
-            <STableCell>
-                <SBadge :color="getRoleColor(row.role)">{{ row.role }}</SBadge>
-            </STableCell>
-            <STableCell class="w-[1%]">
-                <div class="flex h-full gap-2 [&>*]:h-5 [&>*]:w-5">
-                    <button><QrCodeIcon /></button>
-                    <button><DocumentDuplicateIcon /></button>
-                    <button><TrashIcon class="text-red-500" /></button>
-                </div>
             </STableCell>
         </STableRow>
     </STableBody>
 </STable>`,
 });
 
+export const InlineAutoGeneration = createVariation({
+    components: {
+        STable,
+        STableHead,
+        STableBody,
+        STableHeadCell,
+        STableRow,
+        STableCell,
+        SBadge,
+        QrCodeIcon,
+        DocumentDuplicateIcon,
+        TrashIcon,
+    },
+    containerClass: 'w-fit',
+    setup: () => {
+        const getRoleColor = (role: string) => {
+            return role === 'Admin' ? 'yellow' : 'green';
+        }
+
+        return { cols: cols.slice(0, 2), rows: rows.slice(0, 4), getRoleColor };
+    },
+    template: `<STable :cols="cols" :rows="rows.map(r => [r.name, r.email])" />`,
+});
+
 export const Borderless = createVariation({
+    components: {
+        STable,
+        STableHead,
+        STableBody,
+        STableHeadCell,
+        STableRow,
+        STableCell,
+        SBadge,
+        QrCodeIcon,
+        DocumentDuplicateIcon,
+        TrashIcon,
+    },
+    containerClass: 'w-fit',
+    setup: () => {
+        const getRoleColor = (role: string) => {
+            return role === 'Admin' ? 'yellow' : 'green';
+        }
+
+        return { cols: cols.slice(0, 2), rows: rows.slice(0, 4), getRoleColor };
+    },
+    template: `<STable borderless :cols="cols" :rows="rows.map(r => [r.name, r.email])" />`,
+});
+
+export const HighlightColumn = createVariation({
+    components: {
+        STable,
+        STableHead,
+        STableBody,
+        STableHeadCell,
+        STableRow,
+        STableCell,
+        SBadge,
+        QrCodeIcon,
+        DocumentDuplicateIcon,
+        TrashIcon,
+    },
+    containerClass: 'w-fit',
+    setup: () => {
+        const getRoleColor = (role: string) => {
+            return role === 'Admin' ? 'yellow' : 'green';
+        }
+
+        return { cols: cols.slice(0, 3), rows: rows.slice(0, 4), getRoleColor };
+    },
+    template: `<STable :cols="cols">
+    <STableBody>
+        <STableRow v-for="row in rows" :key="row.email">
+            <STableCell highlight>
+                {{ row.name }}
+            </STableCell>
+            <STableCell>
+                {{ row.email }}
+            </STableCell>
+            <STableCell highlight>
+                {{ row.title }}
+            </STableCell>
+        </STableRow>
+    </STableBody>
+</STable>`,
+});
+
+export const HighlightColumnForInlineAutoGenerationMode = createVariation({
+    components: {
+        STable,
+        STableHead,
+        STableBody,
+        STableHeadCell,
+        STableRow,
+        STableCell,
+        SBadge,
+        QrCodeIcon,
+        DocumentDuplicateIcon,
+        TrashIcon,
+    },
+    containerClass: 'w-fit',
+    setup: () => {
+        const getRoleColor = (role: string) => {
+            return role === 'Admin' ? 'yellow' : 'green';
+        }
+
+        return { cols: cols.slice(0, 3), rows: rows.slice(0, 4), getRoleColor };
+    },
+    template: `<STable highlight="[0, 2]" :cols="cols" :rows="rows.map(r => [r.name, r.email, r.title])" />`,
+});
+
+export const Customize = createVariation({
     components: {
         STable,
         STableHead,
@@ -252,7 +349,7 @@ export const Borderless = createVariation({
 
         return { cols, rows, getRoleColor };
     },
-    template: `<STable borderless>
+    template: `<STable>
     <STableHead>
         <STableHeadCell v-for="col in cols" :key="col">{{ col }}</STableHeadCell>
     </STableHead>