+++ src/components/spartan/STable/STableRow.vue
@@ -1,9 +1,13 @@
 <script setup lang="ts">
 import { twMerge } from 'tailwind-merge';
+import { useContext } from './api';
 
 defineProps<{
     class?: string;
 }>();
+
+const context = useContext('STableHeadCell');
+context.register('row');
 </script>
 
 <template>